factor out MenuButtonWidget
