+++ src/admin/ui/config/gameLayout/VerbsUI.tsx
@@ -1,22 +1,10 @@
 import { setSelected } from 'admin/store/reducers/editorStateReducer/selectedEntityReducer';
 import { setVerbPosition } from 'admin/store/reducers/gameStateReducer/configReducer/positionsReducer';
 import { useSelector, useDispatch } from 'admin/ui/hooks/redux';
-import useCachebuster from 'admin/ui/hooks/useCachebuster';
-import { TextStateless as Text, makeCanvasSet } from 'game/ui/shared/Text';
 import { KonvaEventObject } from 'konva/types/Node';
 import React from 'react';
-import { Group, Image } from 'react-konva';
-import { createSelector } from 'reselect';
-import { setCursorStyle } from 'shared/components/PreciseImage';
-
-const canvasesSelector = createSelector(
-  // TODO: fix the "any" type here
-  (state: any) => state.gameState.images,
-  (images) => {
-    const charSet = images['alpha-dark'];
-    return charSet ? makeCanvasSet(charSet) : undefined;
-  },
-);
+import { Group } from 'react-konva';
+import MenuButtonWidget from './MenuButtonWidget';
 
 type Props = {
   top: number;
@@ -27,32 +15,24 @@ const Verb = (props: Props) => {
   const { top, left, index } = props;
   const dispatch = useDispatch();
   const verbs = useSelector(state => state.gameState.config.verbs);
-  const images = useSelector(state => state.gameState.images);
-  const canvases = useSelector(canvasesSelector);
-  const cachebuster = useCachebuster(top + left);
 
   return (
-    <Group
-      x={left + cachebuster}
-      y={top + cachebuster}
-      draggable
+    <MenuButtonWidget
+      left={left}
+      top={top}
+      text={verbs[index].name}
       onClick={() => dispatch(setSelected({
         id: index,
         type: 'verbs',
       }))}
-      onDragEnd={(e: KonvaEventObject<DragEvent>) => {
+      onDrag={(e: KonvaEventObject<DragEvent>) => {
         dispatch(setVerbPosition({
           top: Math.round(e.target.y()),
           left: Math.round(e.target.x()),
           index,
         }));
       }}
-      onMouseEnter={setCursorStyle('pointer')}
-      onMouseLeave={setCursorStyle('default')}
-    >
-      <Image image={images['menu-button']} />
-      <Text text={verbs[index].name} left={9} top={1} canvases={canvases} />
-    </Group>
+    />
   );
 };
